% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ordered_data.R
\name{create_ordered_data}
\alias{create_ordered_data}
\title{Order the data to be written}
\usage{
create_ordered_data(gt_table)
}
\arguments{
\item{gt_table}{A gt table}
}
\value{
dataframe with the columns in correct order or removed.
}
\description{
Order the data to be written
}
\examples{
example_0 <- gt::gtcars |>
  gt::gt()

gtcars_example <- example_0 |>
  create_ordered_data()
}
