% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spanner.R
\name{create_spanner_helper}
\alias{create_spanner_helper}
\title{Create a helper to write spanners}
\usage{
create_spanner_helper(gt_table, ordered_gt_data)
}
\arguments{
\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}
}
\value{
A list of helpers.
}
\description{
It will create a list for information:
- A list for each level spanner.
}
\details{
Within each level list, there will one list for each spanner with label and columns numbers.
}
