% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spanner.R
\name{get_spanner_info}
\alias{get_spanner_info}
\title{Get information for a given spanner}
\usage{
get_spanner_info(gt_table, row, ordered_gt_data)
}
\arguments{
\item{gt_table}{A gt table}

\item{row}{number of the row if the `_spanners` df}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}
}
\value{
a list with the label spanner and columns numbers.
}
\description{
Get information for a given spanner
}
