% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_to_xlsx.R
\name{gt_to_xlsx}
\alias{gt_to_xlsx}
\title{Write a gt table to an xlsx workbook}
\usage{
gt_to_xlsx(gt_table, wb, sheet_name)
}
\arguments{
\item{gt_table}{A gt table}

\item{wb}{Name of Workbook object like `openxlsx::createWorkbook()`}

\item{sheet_name}{name of the sheet in the workbook}
}
\value{
Nothing. It will update the wb object.
}
\description{
Write a gt table to an xlsx workbook
}
\examples{
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "gtcars_example_0")

example_0 <- gt::gtcars |>
  gt::gt()

example_0 |>
  gt_to_xlsx(wb = wb,sheet_name = "gtcars_example_0")
}
