% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtcars_8}
\alias{gtcars_8}
\title{gtcars 8 example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 15 columns.
}
\source{
<https://gt.rstudio.com/articles/case-study-gtcars.html>
}
\usage{
gtcars_8
}
\description{
Example dataset in the vignette Case Study: gtcars from `gt` package.
}
\keyword{datasets}
