% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spanner.R
\name{write_one_spanner}
\alias{write_one_spanner}
\title{Helper to write one spanner}
\usage{
write_one_spanner(
  spanner_info,
  wb,
  sheet_name,
  row_to_start,
  style = column_labels_border
)
}
\arguments{
\item{spanner_info}{helper with spanner columns number and label.}

\item{wb}{Name of Workbook object like `openxlsx::createWorkbook()`}

\item{sheet_name}{name of the sheet in the workbook}

\item{row_to_start}{Number of the row where to start writting the stub and body.}

\item{style}{style to use to format the spanner.}
}
\value{
Nothing. It will update the wb object with one spanner.
}
\description{
It will write one spanner.
}
