% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_source_note.R
\name{write_source_note}
\alias{write_source_note}
\title{Write table headers}
\usage{
write_source_note(wb, row_to_start, gt_table, ordered_gt_data)
}
\arguments{
\item{wb}{Name of the Workbook object created with `openxlsx2::wb_workbook()`}

\item{row_to_start}{Number of the row where to start writing the stub and body.}

\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}
}
\value{
wb with the active sheet with the source note.
}
\description{
Write table headers
}
