% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spanner.R
\name{write_spanners}
\alias{write_spanners}
\title{Write spanners}
\usage{
write_spanners(gt_table, ordered_gt_data, wb, sheet_name, row_to_start)
}
\arguments{
\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}

\item{wb}{Name of Workbook object like `openxlsx::createWorkbook()`}

\item{sheet_name}{name of the sheet in the workbook}

\item{row_to_start}{Number of the row where to start writting the stub and body.}
}
\value{
Nothing. It will update the wb object with all spanners.
}
\description{
It creates a spanner helper with `create_ordered_data` then uses `write_one_spanner` to write all
spanners.
The cells of the spanner will be merged.
}
\examples{
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "one_spanner")

tab_with_spanner <-
  gtcars_8 |>
  gt::gt() |>
  gt::tab_spanner(
    label = "Performance",
    columns = c(mpg_c, mpg_h, hp, hp_rpm, trq, trq_rpm)
  )

ordered_data <- tab_with_spanner |>
  create_ordered_data()

write_spanners(
  gt_table = tab_with_spanner,
  ordered_gt_data = ordered_data,
  wb = wb,
  sheet_name = "one_spanner",
  row_to_start = 1
)
}
