% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stub_and_table_body.R
\name{write_stub_and_table_body}
\alias{write_stub_and_table_body}
\title{Write stub and table body}
\usage{
write_stub_and_table_body(
  gt_table,
  ordered_gt_data,
  wb,
  sheet_name,
  row_to_start
)
}
\arguments{
\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}

\item{wb}{Name of Workbook object like `openxlsx::createWorkbook()`}

\item{sheet_name}{name of the sheet in the workbook}

\item{row_to_start}{Number of the row where to start writting the stub and body.}
}
\value{
It will update the wb object with stub and body table. In addition, it will return a list
with the first row and last row with information for each group.
}
\description{
Write stub and table body
}
\examples{
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "gtcars_example")

example_0 <- gt::gtcars |>
  gt::gt()

gtcars_example <- example_0 |>
  create_ordered_data()

example_0 |>
  write_stub_and_table_body(ordered_gt_data = gtcars_example, wb, "gtcars_example", 1)
}
