% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stub_head_and_column_labels.R
\name{write_stub_head_and_column_labels}
\alias{write_stub_head_and_column_labels}
\title{Write stub and columns label}
\usage{
write_stub_head_and_column_labels(
  gt_table,
  ordered_gt_data,
  wb,
  sheet_name,
  row_to_start
)
}
\arguments{
\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}

\item{wb}{Name of Workbook object like `openxlsx::createWorkbook()`}

\item{sheet_name}{name of the sheet in the workbook}

\item{row_to_start}{Number of the row where to start writting the stub and body.}
}
\value{
Nothing. It will update the wb object with stub and columns labels.
}
\description{
Write stub and columns label
}
\examples{
wb <- openxlsx::createWorkbook()
openxlsx::addWorksheet(wb, "gtcars_example")

ordered_example <- gtopenxlsx::gtcars_8 |>
  gt::gt() |>
  create_ordered_data()
gt_table_example <- gtopenxlsx::gtcars_8 |>
  gt::gt()
ordered_example <- gt_table_example |>
  create_ordered_data()
gt_table_example |>
  write_stub_head_and_column_labels(ordered_gt_data = ordered_example, wb, "gtcars_example", 1)
}
