% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_to_xlsx.R
\name{gt_to_xlsx}
\alias{gt_to_xlsx}
\title{Write a gt table to an xlsx workbook}
\usage{
gt_to_xlsx(wb, gt_table)
}
\arguments{
\item{wb}{Name of the Workbook object created with `openxlsx2::wb_workbook()`}

\item{gt_table}{A gt table}
}
\value{
wb with the active sheet with the gt table rendered.
}
\description{
Write a gt table to an xlsx workbook
}
\examples{
wb <- openxlsx2::wb_workbook() |>
  openxlsx2::wb_add_worksheet("gtcars_example_0")
example_0 <- gt::gtcars |>
  gt::gt()
wb <- gt_to_xlsx(wb = wb,example_0)

}
