% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spanner.R
\name{write_one_spanner}
\alias{write_one_spanner}
\title{Helper to write one spanner}
\usage{
write_one_spanner(wb, spanner_info, row_to_start)
}
\arguments{
\item{wb}{Name of the Workbook object created with `openxlsx2::wb_workbook()`}

\item{spanner_info}{helper with spanner columns number and label.}

\item{row_to_start}{Number of the row where to start writing the stub and body.}

\item{style}{style to use to format the spanner.}
}
\value{
Nothing. It will update the wb object with one spanner.
}
\description{
It will write one spanner.
}
