% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_spanner.R
\name{write_spanners}
\alias{write_spanners}
\title{Write spanners}
\usage{
write_spanners(wb, gt_table, ordered_gt_data, row_to_start)
}
\arguments{
\item{wb}{Name of the Workbook object created with `openxlsx2::wb_workbook()`}

\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}

\item{row_to_start}{Number of the row where to start writing the stub and body.}
}
\value{
wb with the active sheet with all spanners.
}
\description{
It creates a spanner helper with `create_ordered_data` then uses `write_one_spanner` to write all
spanners.
The cells of the spanner will be merged.
}
\examples{
wb <- openxlsx2::wb_workbook() |>
  openxlsx2::wb_add_worksheet()
tab_with_spanner <-
  gtcars_8 |>
  gt::gt() |>
  gt::tab_spanner(
    label = "Performance",
    columns = c(mpg_c, mpg_h, hp, hp_rpm, trq, trq_rpm)
  )
ordered_data <- tab_with_spanner |>
  create_ordered_data()
wb <- wb |> write_spanners(
  gt_table = tab_with_spanner,
  ordered_gt_data = ordered_data,
  row_to_start = 1
)
}
