% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stub_and_table_body.R
\name{write_stub_and_table_body}
\alias{write_stub_and_table_body}
\title{Write stub and table body}
\usage{
write_stub_and_table_body(wb, row_to_start, gt_table, ordered_gt_data)
}
\arguments{
\item{wb}{Name of the Workbook object created with `openxlsx2::wb_workbook()`}

\item{row_to_start}{Number of the row where to start writing the stub and body.}

\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}
}
\value{
Returns a list with :
   - wb with the active sheet with stub and body table.
   - a list with the first row and last row with information for each group.
}
\description{
Write stub and table body
}
\examples{
wb <- openxlsx2::wb_workbook() |>
  openxlsx2::openxlsx2::wb_add_worksheet()

example_0 <- gt::gtcars |>
  gt::gt()

gtcars_example <- example_0 |>
  create_ordered_data()

wb <- example_0 |>
  write_stub_and_table_body(ordered_gt_data = gtcars_example,1)
}
