% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_stub_head_and_column_labels.R
\name{write_stub_head_and_column_labels}
\alias{write_stub_head_and_column_labels}
\title{Write stub and columns label}
\usage{
write_stub_head_and_column_labels(wb, row_to_start, gt_table, ordered_gt_data)
}
\arguments{
\item{wb}{Name of the Workbook object created with `openxlsx2::wb_workbook()`}

\item{row_to_start}{Number of the row where to start writing the stub and body.}

\item{gt_table}{A gt table}

\item{ordered_gt_data}{ordered data from `create_ordered_data`}
}
\value{
wb with the active sheet with stub and columns labels.
}
\description{
Write stub and columns label
}
\examples{
wb <- openxlsx2::wb_workbook() |>
  openxlsx2::wb_add_worksheet()

gt_table_example <- gtcars_8 |>
  gt::gt()
ordered_example <- gt_table_example |>
  create_ordered_data()
wb <- wb |>
  write_stub_head_and_column_labels(row_to_start = 1,
                                    gt_table = gt_table_example,
                                    ordered_gt_data = ordered_example)

}
